Make FileAvatar & AsyncAvatar that takes bytes
We have been repeating a set of widgets to represent loading bytes.
Here we have combined the widgets into one.
